refactor(approve): relocate email bits to match feature-scoped layout
Keeping CLAUDE.md's rule honest — lib shouldn't depend on React, and
approve-specific logic shouldn't pretend to be cross-cutting.

- emails/approve/{signer-invited,document-completed}.tsx
  (React Email templates live in their own top-level folder; matches the
  community convention and opens the door to `react-email dev` preview)
- lib/approve/email-drain.ts (was lib/email/drain-outbox.ts — it only
  knows approve, so it belongs with lib/approve/*)
- lib/email/resend.ts stays as the one actual cross-app helper
